Refactoring colony integration
Closes #9
